pyproject.toml
[project]
name = "crypto-currency-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nest-asyncio>=1.6.0",
    "openai-agents>=0.2.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.47.0",
]
